create RescaleEvent stub
