+++ test/cement-layer.test.ts
@@ -1,6 +1,5 @@
-import { scaleLinear } from 'd3-scale';
-import { zoomIdentity } from 'd3-zoom';
 import { CementLayer, IntersectionReferenceSystem } from '../src/index';
+import { rescaleEventStub } from './test-helpers';
 
 describe('CementLayer', () => {
   let elm: HTMLElement;
@@ -23,21 +22,6 @@ describe('CementLayer', () => {
       cement: [{ casingIds: ['1'], toc: 250 }],
       holes: [{ start: 50, end: 250, diameter: 36 }],
     };
-    const xBounds = [0, 1000] as [number, number];
-    const yBounds = [0, 1000] as [number, number];
-    const layerEvent = {
-      xBounds,
-      yBounds,
-      zFactor: 1,
-      viewportRatio: 1,
-      xRatio: 1,
-      yRatio: 1,
-      width: 1,
-      height: 1,
-      transform: zoomIdentity,
-      xScale: scaleLinear().domain(xBounds).range([0, 1]),
-      yScale: scaleLinear().domain(yBounds).range([0, 1]),
-    };
 
     it('should render when reference system is set in constructor', () => {
       // Arrange
@@ -45,7 +29,7 @@ describe('CementLayer', () => {
       const layer = new CementLayer('casing-layer', { referenceSystem });
       layer.onMount({ elm });
       layer.onUpdate({});
-      layer.onRescale(layerEvent);
+      layer.onRescale(rescaleEventStub(data));
 
       // Act
       layer.data = data;
@@ -61,7 +45,7 @@ describe('CementLayer', () => {
       layer.referenceSystem = referenceSystem;
       layer.onMount({ elm });
       layer.onUpdate({});
-      layer.onRescale(layerEvent);
+      layer.onRescale(rescaleEventStub(data));
 
       // Act
       layer.data = data;
@@ -75,7 +59,7 @@ describe('CementLayer', () => {
       const layer = new CementLayer('casing-layer', {});
       layer.onMount({ elm });
       layer.onUpdate({});
-      layer.onRescale(layerEvent);
+      layer.onRescale(rescaleEventStub(data));
 
       // Act
       // Assert
